copy README.md, LICENSE.md to src/site/markdown/ to update files
it might have worked to just take a look at the logically linked file, but now I have removed that and made the work I did before.
Also, as I have discovered: the links works somewhat on Windows:
it does display the file but updates upon changes of the original
are not transferred over the linked one.